18-1.1.0 Part  1 Частичная имплементация ORM, подготовка к распараллеливанию запросов к серверу и записи в БД

+++ Dastarhan/app/src/main/java/mobi/esys/dastarhan/database/CuisineRealmRepository.java
@@ -29,7 +29,7 @@ class CuisineRealmRepository implements CuisineRepository {
     }
 
     @Override
-    public Cuisine getById(final long id) {
+    public Cuisine getById(final int id) {
         return realmTemplate.findInRealm(new RealmTransactionCallback<Cuisine>() {
             @Override
             public Cuisine execute(Realm realm) {
