+++ Dastarhan/app/src/main/java/mobi/esys/dastarhan/database/CuisineRepository.java
@@ -6,7 +6,7 @@ public interface CuisineRepository {
 
     void addOrUpdate(Cuisine cuisine);
 
-    Cuisine getById(long id);
+    Cuisine getById(int id);
 
     List<Cuisine> getAll();
 }
