+++ Dastarhan/app/src/main/java/mobi/esys/dastarhan/database/PromoRepository.java
@@ -6,7 +6,7 @@ public interface PromoRepository {
 
     void addOrUpdate(Promo promo);
 
-    Promo getById(long id);
+    Promo getById(int id);
 
     List<Promo> getAll();
 }
